Reduce the number of allocations (and consequently CPU time) performed by the `correlateEndpoints` function removing a duplicated DeepCopy of all endpoints. Indeed, GetEndpoints already returns a DeepCopy, and we can subsequently mutate the endpoint information in-place, without needing to copy them again. Additionally, let's rework a bit the logic
to avoid unnecessarily creating a second map and copy all entries there. 

Considering `BenchmarkCorrelateEndpoints` which calls `correlateEndpoints` for a service associated with 10 epslices, each with 100 endpoints (2 ports each), these changes lead to approximately a 50% CPU and memory allocations reduction:

```
goos: linux
goarch: amd64
pkg: github.com/cilium/cilium/pkg/k8s
cpu: 13th Gen Intel(R) Core(TM) i7-13800H
                      │   old.txt    │               new.txt               │
                      │    sec/op    │   sec/op     vs base                │
CorrelateEndpoints-20   1140.8µ ± 4%   605.0µ ± 8%  -46.96% (p=0.000 n=10)

                      │    old.txt    │               new.txt                │
                      │     B/op      │     B/op      vs base                │
CorrelateEndpoints-20   1177.9Ki ± 0%   589.0Ki ± 0%  -50.00% (p=0.000 n=10)

                      │   old.txt    │               new.txt               │
                      │  allocs/op   │  allocs/op   vs base                │
CorrelateEndpoints-20   10.062k ± 0%   5.031k ± 0%  -50.00% (p=0.000 n=10)
```