@giorio94 2024-10-30
Apparently kicked out of the merge queue due to the ongoing GitHub Actions incident: https://www.githubstatus.com/